pyproject.toml
[project]
name = "urlshortner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "pillow>=11.3.0",
    "qrcode>=8.2",
]
